requirements.txt
ipython
pillow
numpy
torch==1.12.0
requests
tqdm
matplotlib

